backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
httpx==0.25.1
pydantic==2.5.0
pyyaml==6.0.1
langchain==0.1.0
langchain-openai==0.0.5
openai==1.10.0
PyJWT==2.8.0